auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
email-validator==2.1.1
bcrypt==3.2.2